# optimization: enqueue transactions in parallel from p2p